Changed id from Long to Integer

+++ src/main/java/com/book/list/model/Book.java
@@ -2,20 +2,20 @@ package com.book.list.model;
 
 public class Book {
 
-    private Long id;
+    private Integer id;
     private String title;
     private String author;
 
     public Book() {
     }
 
-    public Book(Long id, String title, String author) {
+    public Book(Integer id, String title, String author) {
         this.id = id;
         this.title = title;
         this.author = author;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -27,7 +27,7 @@ public class Book {
         return author;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 }
